A Star Default
The default implementation of a star has been coded in.

There is also now new file loading for a specific txt file. I am using my previous cmp105 project to build the worlds, and I am exporting them in text form. This will allow me to easily create different types of nodemaps.

I have come to realise that it's a bit messy and I should have really stuck with a 1x1 movement grid for simplicity. In practical terms if there were different size nodes, I could have a variable for them and let the game itself multiply the movement. TODO <-;

TODO: Seperate the A star and Node files into their own version. Although it would be nice to have it all in one class, I don't believe it is at all possible to do it with the nodemaps, as the nodes will require different amounts of information depending on what implementation of a star it is using - great way to see differences in performance.

Todo: Put the file loading into a class, don't know which one but I would rather compartmentalise it instead of keeping it in source.cpp